% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_plot.R
\name{binomial_plot}
\alias{binomial_plot}
\title{Binomial plot}
\usage{
binomial_plot(n, p)
}
\arguments{
\item{n}{Number of trials}

\item{p}{success probability}
}
\value{
ggplot2 object
}
\description{
Plots a binomial distribution using ggplot2
}
\examples{
binomial_plot(10, .5)
}
