% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_plot.R
\name{exp_plot}
\alias{exp_plot}
\title{Exponential distribution plot}
\usage{
exp_plot(r, max_x = 25, probs = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{r}{rate parameter}

\item{max_x}{max value x axis}

\item{probs}{quantiles to indicate in the plot}
}
\value{
ggplot2 object
}
\description{
Exponential distribution plot
}
\examples{
exp_plot(0.1)
}
