% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norm_densities.R
\name{get_norm_density}
\alias{get_norm_density}
\title{Get normal densities, shifted and scaled}
\usage{
get_norm_density(
  outcome_mean = 0,
  error_sigma = 1,
  dens_spread_factor = 1,
  predictor_value = 0,
  z = 2
)
}
\arguments{
\item{outcome_mean}{the mean value of the normal distribution}

\item{error_sigma}{the spead (sd) of the normal distribution}

\item{dens_spread_factor}{spread factor of the densities}

\item{predictor_value}{offset to be added to the densities}

\item{z}{z-value indicating start and end in z-values of the densities}
}
\description{
Gives densities of a specified normal distribution.
}
\examples{
lm1 <- lm(mpg ~ hp, data = mtcars)
lm1_pred_200 <- predict(lm1, newdata = tibble(hp = 200))

ggplot(mtcars) +
  aes(x = hp, y = mpg) +
  geom_point() +
  geom_abline(slope = coef(lm1)[2], intercept = coef(lm1)[1], color = "blue") +
  geom_path(
    data = get_norm_density(
      outcome_mean = lm1_pred_200,
      error_sigma = 2,
      z = 3,
      dens_spread_factor = 5e1,
      predictor_value = 200
    ),
    aes(y = outcome, x = norm_dens)
  )
}
