% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shaded_normal.R
\name{shaded_normal}
\alias{shaded_normal}
\title{Plot shaded normal}
\usage{
shaded_normal(
  xmin = -3,
  xmax = 3,
  threshold_min = -1.96,
  threshold_max = 1.96,
  color1 = "#E69F00",
  color2 = "#56B4E9",
  alpha_value = 0.9,
  breaks = c(threshold_min, threshold_max)
)
}
\arguments{
\item{xmin}{min value of the the plotted normal (dbl)}

\item{xmax}{max value of the the plotted normal (dbl)}

\item{threshold_min}{lower threshold (default -1.96)}

\item{threshold_max}{upper threshold (default 1.96)}

\item{color1}{fill color of central part of the distribution (#E69F00)}

\item{color2}{fill color of extremal part of the distribution (#56B4E9)}

\item{alpha_value}{alpha of the fill (default .9)}
}
\value{
ggplot object
}
\description{
plots a standard normal distribution, where a lower/upper part
(2.5\%, 97.5\% per default) is filled with some other color
}
\examples{
shaded_normal() + labs(x = "", y = "")
shaded_normal(breaks = -3:3) + labs(x = "", y = "")
shaded_normal(breaks = -3:3) +
  theme(panel.grid.major.x = element_blank(),
  panel.grid.minor.x = element_blank())
}
