% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacherplot-package.R
\docType{package}
\name{teacherplot-package}
\alias{teacherplot}
\alias{teacherplot-package}
\title{teacherplot: Providing Some Plots Teachers May Like.}
\description{
Some typical graphs potentially useful for teaching statistics and related fields.
}
\keyword{internal}
