% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_plot.R
\name{uniform_plot}
\alias{uniform_plot}
\title{Uniform plot}
\usage{
uniform_plot(a, b)
}
\arguments{
\item{a}{minimum value of the uniform distribution}

\item{b}{maximum value of the uniform distribution}
}
\value{
ggplot2 object
}
\description{
Plots a uniform distribution with parameters a and b, using ggplot2
}
\examples{
uniform_plot(0, 1)
}
