% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacherplot-package.R
\docType{package}
\name{teacherplot-package}
\alias{teacherplot}
\alias{teacherplot-package}
\title{teacherplot: Providing Some Plots Teachers May Like.}
\description{
Some typical graphs potentially useful for teaching statistics and related fields.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebastiansauer/teacherplot}
  \item \url{https://sebastiansauer.github.io/teacherplot/}
  \item Report bugs at \url{https://github.com/sebastiansauer/teacherplot/issues}
}

}
\keyword{internal}
